parser: new explicit diagnostic if keyword a start of stmt is unknown
This specifically triggers for a statement like the one below:

 DOP TABLE table;

'DOP' is obviously missspelled and the user intendend to use the 'DROP' keyword, the parser now detects this specific case and generates a diagnostic. This diagnostic will include a suggestion for similar keywords in the future, once the lev::distance function is implemented.